+++ sdk/src/driftClient.ts
@@ -1114,7 +1114,7 @@ export class DriftClient {
 			await this.program.instruction.initializeSignedMsgUserOrders(numOrders, {
 				accounts: {
 					signedMsgUserOrders: signedMsgUserAccountPublicKey,
-					authority: this.wallet.publicKey,
+					authority,
 					payer: this.wallet.publicKey,
 					rent: anchor.web3.SYSVAR_RENT_PUBKEY,
 					systemProgram: anchor.web3.SystemProgram.programId,
@@ -1127,10 +1127,15 @@ export class DriftClient {
 	public async resizeSignedMsgUserOrders(
 		authority: PublicKey,
 		numOrders: number,
+		userSubaccountId?: number,
 		txParams?: TxParams
 	): Promise<TransactionSignature> {
 		const resizeUserAccountIx =
-			await this.getResizeSignedMsgUserOrdersInstruction(authority, numOrders);
+			await this.getResizeSignedMsgUserOrdersInstruction(
+				authority,
+				numOrders,
+				userSubaccountId
+			);
 		const tx = await this.buildTransaction([resizeUserAccountIx], txParams);
 		const { txSig } = await this.sendTransaction(tx, [], this.opts);
 
@@ -1139,7 +1144,8 @@ export class DriftClient {
 
 	async getResizeSignedMsgUserOrdersInstruction(
 		authority: PublicKey,
-		numOrders: number
+		numOrders: number,
+		userSubaccountId?: number
 	): Promise<TransactionInstruction> {
 		const signedMsgUserAccountPublicKey = getSignedMsgUserAccountPublicKey(
 			this.program.programId,
@@ -1149,8 +1155,14 @@ export class DriftClient {
 			await this.program.instruction.resizeSignedMsgUserOrders(numOrders, {
 				accounts: {
 					signedMsgUserOrders: signedMsgUserAccountPublicKey,
-					authority: this.wallet.publicKey,
+					authority,
+					payer: this.wallet.publicKey,
 					systemProgram: anchor.web3.SystemProgram.programId,
+					user: await getUserAccountPublicKey(
+						this.program.programId,
+						authority,
+						userSubaccountId
+					),
 				},
 			});
 
@@ -2067,16 +2079,18 @@ export class DriftClient {
 	 * @param subAccountId
 	 */
 	public async forceGetUserAccount(
-		subAccountId?: number
+		subAccountId?: number,
+		authority?: PublicKey
 	): Promise<UserAccount | undefined> {
-		await this.getUser(subAccountId).fetchAccounts();
-		return this.getUser(subAccountId).getUserAccount();
+		await this.getUser(subAccountId, authority).fetchAccounts();
+		return this.getUser(subAccountId, authority).getUserAccount();
 	}
 
 	public getUserAccountAndSlot(
-		subAccountId?: number
+		subAccountId?: number,
+		authority?: PublicKey
 	): DataAndSlot<UserAccount> | undefined {
-		return this.getUser(subAccountId).getUserAccountAndSlot();
+		return this.getUser(subAccountId, authority).getUserAccountAndSlot();
 	}
 
 	public getSpotPosition(
@@ -2175,7 +2189,14 @@ export class DriftClient {
 			this.getRemainingAccountMapsForUsers(params.userAccounts);
 
 		if (params.useMarketLastSlotCache) {
-			const lastUserSlot = this.getUserAccountAndSlot()?.slot;
+			const lastUserSlot = this.getUserAccountAndSlot(
+				params.userAccounts.length > 0
+					? params.userAccounts[0].subAccountId
+					: this.activeSubAccountId,
+				params.userAccounts.length > 0
+					? params.userAccounts[0].authority
+					: this.authority
+			)?.slot;
 
 			for (const [
 				marketIndex,
@@ -2583,7 +2604,7 @@ export class DriftClient {
 		let remainingAccounts = [];
 		if (userInitialized) {
 			remainingAccounts = this.getRemainingAccounts({
-				userAccounts: [await this.forceGetUserAccount()],
+				userAccounts: [await this.forceGetUserAccount(subAccountId)],
 				useMarketLastSlotCache: true,
 				writableSpotMarketIndexes: [marketIndex],
 			});
@@ -6392,7 +6413,7 @@ export class DriftClient {
 	): Promise<TransactionInstruction[]> {
 		const remainingAccounts = this.getRemainingAccounts({
 			userAccounts: [takerInfo.takerUserAccount],
-			useMarketLastSlotCache: true,
+			useMarketLastSlotCache: false,
 			readablePerpMarketIndex: marketIndex,
 		});
 
@@ -6510,7 +6531,7 @@ export class DriftClient {
 				this.getUserAccount(subAccountId),
 				takerInfo.takerUserAccount,
 			],
-			useMarketLastSlotCache: true,
+			useMarketLastSlotCache: false,
 			writablePerpMarketIndexes: [orderParams.marketIndex],
 		});
 
@@ -9289,6 +9310,37 @@ export class DriftClient {
 		return [verifyIx, ix];
 	}
 
+	public async getPostManySwitchboardOnDemandUpdatesAtomicIxs(
+		feeds: PublicKey[],
+		recentSlothash?: Slothash,
+		numSignatures = 3
+	): Promise<TransactionInstruction[] | undefined> {
+		const program = await this.getSwitchboardOnDemandProgram();
+		for (const feed of feeds) {
+			const feedAccount = new PullFeed(program, feed);
+			if (!this.sbProgramFeedConfigs) {
+				this.sbProgramFeedConfigs = new Map();
+			}
+			if (!this.sbProgramFeedConfigs.has(feedAccount.pubkey.toString())) {
+				const feedConfig = await feedAccount.loadConfigs();
+				this.sbProgramFeedConfigs.set(feed.toString(), feedConfig);
+			}
+		}
+
+		const [pullIxs, _responses, success] =
+			await PullFeed.fetchUpdateManyLightIx(program, {
+				feeds,
+				numSignatures,
+				recentSlothashes: recentSlothash
+					? [[new BN(recentSlothash.slot), recentSlothash.hash]]
+					: undefined,
+			});
+		if (!success) {
+			return undefined;
+		}
+		return pullIxs;
+	}
+
 	public async getPostSwitchboardOnDemandUpdateAtomicIx(
 		feed: PublicKey,
 		recentSlothash?: Slothash,
@@ -9303,18 +9355,20 @@ export class DriftClient {
 			const feedConfig = await feedAccount.loadConfigs();
 			this.sbProgramFeedConfigs.set(feed.toString(), feedConfig);
 		}
-		const [pullIx, _responses, success] = await feedAccount.fetchUpdateIx(
+		const [pullIx, _responses, success] = await PullFeed.fetchUpdateManyIx(
+			program,
 			{
+				feeds: [feed],
 				numSignatures,
-			},
-			recentSlothash
-				? [[new BN(recentSlothash.slot), recentSlothash.hash]]
-				: undefined
+				recentSlothashes: recentSlothash
+					? [[new BN(recentSlothash.slot), recentSlothash.hash]]
+					: undefined,
+			}
 		);
 		if (!success) {
 			return undefined;
 		}
-		return pullIx;
+		return pullIx[0];
 	}
 
 	public async postSwitchboardOnDemandUpdate(